social_media_api/requirements.txt
Django==4.2.28
djangorestframework==3.14.0
Pillow==10.0.0
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==2.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
